refactor: 불필요 파일 제거
