--- src/hooks/query/useCartItemsQuery.ts
@@ -1,9 +1,0 @@
-import { CartItemApi } from "@/apis";
-import { useQuery } from "@/modules";
-
-export default function useCartItemsQuery() {
-  return useQuery({
-    queryKey: "cartItems",
-    queryFn: CartItemApi.getCartItems,
-  });
-}
