--- src/hooks/useAutoFocus/index.ts
@@ -1,1 +1,0 @@
-export { default as useAutoFocus } from './useAutoFocus';
