+++ src/hooks/useDevice/index.ts
@@ -1,3 +1,2 @@
-export { default as useDevice } from './useDevice';
-export * from './type';
-export * from './constant';
+export * from "./type";
+export * from "./constant";
